+++ addOns/pscanrulesBeta/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Fixed
 - PII Disclosure scan rule now ignores images (Issue 6697).
+- PII Disclosure scan rule will now ignore seeming decimal numbers unless at Low threshold (Issue 6639).
 
 ## [25] - 2021-06-17
 ### Changed